Move mvsData/Color.hpp to image module
ColorRGBf and ColorRGBAf are practically identical to RGBfColor and
RGBAfColor. Eventually we should merge the classes and have a single way
to represent a color.
