+++ src/aliceVision/image/CMakeLists.txt
@@ -3,6 +3,7 @@ add_definitions(-DTHIS_SOURCE_DIR="${CMAKE_CURRENT_SOURCE_DIR}")
 # Headers
 set(image_files_headers
   all.hpp
+  Color.hpp
   Image.hpp
   concat.hpp
   convertion.hpp
