+++ src/aliceVision/image/Color.hpp
