+++ src/aliceVision/mesh/Texturing.cpp
@@ -10,8 +10,8 @@
 
 #include <aliceVision/system/Logger.hpp>
 #include <aliceVision/system/MemoryInfo.hpp>
+#include <aliceVision/image/Color.hpp>
 #include <aliceVision/numeric/numeric.hpp>
-#include <aliceVision/mvsData/Color.hpp>
 #include <aliceVision/mvsData/geometry.hpp>
 #include <aliceVision/mvsData/Pixel.hpp>
 #include <aliceVision/mvsData/Image.hpp>
