+++ src/aliceVision/mvsData/imageAlgo.cpp
@@ -1,7 +1,7 @@
 #include "imageAlgo.hpp"
 
 
-#include <aliceVision/mvsData/Color.hpp>
+#include <aliceVision/image/Color.hpp>
 #include <aliceVision/mvsData/Rgb.hpp>
 #include <aliceVision/mvsData/Image.hpp>
